Renamed io package inside knowledge_linker to inout and fixed imports accordingly. Also, fixed issues related to https://github.com/glciampaglia/knowledge_linker/issues/5
