+++ knowledge_linker/inout/ntriples.py
